+++ crates/oxc_linter/src/rules/jest/consistent_test_it.rs
@@ -117,6 +117,14 @@ fn test() {
             "describe(\"suite\", () => { test(\"foo\") })",
             Some(serde_json::json!([{ "withinDescribe": "test" }])),
         ),
+        (
+            "
+                describe('suite', () => {});
+
+                test('foo', () => {});
+            ",
+            Some(serde_json::json!([{ "withinDescribe": "it" }])),
+        ),
     ];
 
     let fail = vec![