+++ crates/oxc_linter/src/rules/vue/require_render_return.rs
@@ -59,7 +59,7 @@ declare_oxc_lint!(
     RequireRenderReturn,
     vue,
     correctness,
-    version = "next",
+    version = "1.67.0",
 );
 
 impl Rule for RequireRenderReturn {