+++ crates/oxc_linter/src/rules/vue/require_slots_as_functions.rs
@@ -60,7 +60,7 @@ declare_oxc_lint!(
     RequireSlotsAsFunctions,
     vue,
     correctness,
-    version = "next",
+    version = "1.67.0",
 );
 
 impl Rule for RequireSlotsAsFunctions {