+++ crates/oxc_linter/src/utils/mod.rs
@@ -25,6 +25,7 @@ mod promise;
 mod react;
 mod react_perf;
 mod regex;
+mod static_value;
 mod this_expression;
 mod typescript;
 mod unicorn;
@@ -34,8 +35,8 @@ mod vue;
 
 pub use self::{
     comment::*, config::*, control_flow::*, express::*, jest::*, jsdoc::*, nextjs::*, promise::*,
-    react::*, react_perf::*, regex::*, this_expression::*, typescript::*, unicorn::*, url::*,
-    vitest::*, vue::*,
+    react::*, react_perf::*, regex::*, static_value::*, this_expression::*, typescript::*,
+    unicorn::*, url::*, vitest::*, vue::*,
 };
 
 /// List of Eslint rules that have TypeScript equivalents.